lab2/lab2/requirements.txt
click==8.1.3
contourpy==1.0.7
cv2-rgb==1.0.0
cycler==0.11.0
Flask==2.2.3
fonttools==4.39.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
render==1.0.0
six==1.16.0
Werkzeug==2.2.3
